Fix comparison between signed and unsigned integer expressions
../../torchaudio/csrc/pybind/sox/effects.cpp:59:20: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     return (bufsiz > kDefaultCapacityInBytes) ? bufsiz
             ~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~

../../torchaudio/csrc/pybind/sox/io.cpp:47:20: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]
     return (bufsiz > kDefaultCapacityInBytes) ? bufsiz
             ~~~~~~~^~~~~~~~~~~~~~~~~~~~~~~~~

+++ torchaudio/csrc/pybind/sox/effects.cpp
@@ -55,7 +55,7 @@ std::tuple<torch::Tensor, int64_t> apply_effects_fileobj(
     // end up retrieving the static variable defined in `_torchaudio`, which is
     // not correct.
     const auto bufsiz = get_buffer_size();
-    const size_t kDefaultCapacityInBytes = 256;
+    const int64_t kDefaultCapacityInBytes = 256;
     return (bufsiz > kDefaultCapacityInBytes) ? bufsiz
                                               : kDefaultCapacityInBytes;
   }();
